+++ src/nl/mrwouter/minetopiafarms/utils/CustomFlags.java
@@ -1,4 +1,4 @@
-package nl.wouter.minetopiafarms.utils;
+package nl.mrwouter.minetopiafarms.utils;
 
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
