+++ src/nl/mrwouter/minetopiafarms/utils/Updat3r.java
@@ -1,4 +1,4 @@
-package nl.wouter.minetopiafarms.utils;
+package nl.mrwouter.minetopiafarms.utils;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -16,7 +16,7 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 
-import nl.wouter.minetopiafarms.Main;
+import nl.mrwouter.minetopiafarms.Main;
 
 public class Updat3r {
