+++ src/nl/mrwouter/minetopiafarms/utils/WorldGuardLegacyManager.java
@@ -1,4 +1,4 @@
-package nl.wouter.minetopiafarms.utils;
+package nl.mrwouter.minetopiafarms.utils;
 
 import java.lang.reflect.Method;
 import java.util.ArrayList;
