+++ src/nl/mrwouter/minetopiafarms/utils/XMaterial.java
@@ -1,4 +1,4 @@
-package nl.wouter.minetopiafarms.utils;
+package nl.mrwouter.minetopiafarms.utils;
 
 import java.util.HashMap;
